Improvements
- Improved exception handling and messaging
- Added log file checking
- Added logging throughout the application
- Added exit codes
- Deny execution by systems other than Windows 10 and 11
- Application checks if an instance is already running and shows an error message if so, closing the second instance

+++ GarbageCleaner.cs
@@ -1,33 +1,37 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using ConstantsDLL;
 using HardwareInfoDLL;
+using LogGeneratorDLL;
 
 namespace OfflineDriverInstallerOOBE
 {
     internal static class GarbageCleaner
     {
-        public static void cleanDirectories(string path, bool verbose)
+        public static void cleanDirectories(string path, LogGenerator log)
         {
-            if (verbose)
-            {
-                Console.WriteLine(StringsAndConstants.ERASING_GARBAGE);
-                Console.WriteLine();
-            }
             string model = HardwareInfo.GetModel();
             if (model == StringsAndConstants.ToBeFilledByOEM || model == "")
                 model = HardwareInfo.GetModelAlt();
-            string p = path + HardwareInfo.GetBIOSType() + "\\" + HardwareInfo.getOSVersion() + "\\" + HardwareInfo.getOSArchAlt() + "\\";
-            DirectoryInfo directory = new DirectoryInfo(p);
-            foreach (DirectoryInfo dir in directory.GetDirectories())
+            string pathExt = path + HardwareInfo.GetBIOSType() + "\\" + HardwareInfo.getOSVersion() + "\\" + HardwareInfo.getOSArchAlt() + "\\";
+            DirectoryInfo directory = new DirectoryInfo(pathExt);
+
+            try
             {
-                if(dir.Name != model)
-                    dir.Delete(true);
+                if (!Directory.Exists(pathExt))
+                    throw new DirectoryNotFoundException(StringsAndConstants.DIRECTORY_DO_NOT_EXIST);
+
+                log.LogWrite(StringsAndConstants.LOG_INFO, StringsAndConstants.ERASING_GARBAGE, string.Empty, StringsAndConstants.consoleOutCLI);
+
+                foreach (DirectoryInfo dir in directory.GetDirectories())
+                {
+                    if (dir.Name != model)
+                        dir.Delete(true);
+                }
+                log.LogWrite(StringsAndConstants.LOG_INFO, StringsAndConstants.ERASING_SUCCESSFUL, string.Empty, StringsAndConstants.consoleOutCLI);
             }
-            if (verbose)
+            catch(DirectoryNotFoundException e)
             {
-                Console.WriteLine(StringsAndConstants.ERASING_SUCCESSFUL);
-                Console.WriteLine();
+                log.LogWrite(StringsAndConstants.LOG_ERROR, pathExt, e.Message, StringsAndConstants.consoleOutCLI);
             }
         }
     }
